1. RestAPI/requirements.txt
# python -m pip install --upgrade pip
# pip install -r requirements.txt    
google-cloud-vision
requests 
python-dotenv
openai